worker/tests/mock_custom/pyproject.toml
[project]
name = "netboxlabs-custom-worker"
version = "0.0.1"  # Overwritten during the build process
description = "NetBox Labs, Worker mock implementation for testing pourposes"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]
maintainers = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

dependencies = [
    "netboxlabs-worker~=0.0.1",
    "netboxlabs-diode-sdk~=0.4",
]

[project.optional-dependencies]
dev = ["black", "check-manifest", "ruff"]
test = ["coverage", "pytest", "pytest-cov"]

[project.urls]
"Homepage" = "https://netboxlabs.com/"

[tool.setuptools]
packages = [
    "nbl_custom",
]
package-data = {"nbl_custom" = ["**/*"]}

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 140

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "R", "UP", "W"]
ignore = ["F401", "D203", "D212", "D400", "D401", "D404", "RET504"]